% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret-expressions.r
\name{assign_variable}
\alias{assign_variable}
\title{Assign variable in locals table}
\usage{
assign_variable(name, value, x, interpreter)
}
\arguments{
\item{name}{token}

\item{value}{value to assign}

\item{x}{expression}

\item{interpreter}{interpreter}
}
\description{
Assign variable in locals table
}
