% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{enumerate_tree}
\alias{enumerate_tree}
\title{Descent the tree and enumerate each node}
\usage{
enumerate_tree(tree, id = 1L)
}
\arguments{
\item{tree}{abstract syntax tree from \code{parse_tokens}}

\item{id}{initial id}
}
\description{
Descent the tree and enumerate each node
}
