% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.r
\name{env_new}
\alias{env_new}
\alias{env_get}
\alias{env_define}
\alias{env_has}
\alias{env_assign}
\title{Create new Lox environment}
\usage{
env_new(parent = emptyenv())

env_get(env, token)

env_define(env, name, value)

env_has(env, name)

env_assign(env, token, value)
}
\arguments{
\item{parent}{parent environment}

\item{env}{lox environment}

\item{token}{token}

\item{name}{string}

\item{value}{value to assign to token}
}
\description{
Create new Lox environment
}
