% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.r
\name{expr}
\alias{expr}
\alias{expr_assignment}
\alias{expr_binary}
\alias{expr_call}
\alias{expr_grouping}
\alias{expr_literal}
\alias{expr_logical}
\alias{expr_unary}
\alias{expr_variable}
\title{Expressions}
\usage{
expr(x, subclass)

expr_assignment(name, value)

expr_binary(left, operator, right)

expr_call(callee, paren, arguments)

expr_grouping(expression)

expr_literal(value)

expr_logical(left, operator, right)

expr_unary(operator, right)

expr_variable(name)
}
\arguments{
\item{name}{token}

\item{value}{expression}

\item{left, right}{expression}

\item{operator}{token}

\item{callee}{expression}

\item{paren}{token}

\item{arguments}{list of expressions}
}
\description{
Expressions
}
