% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.r
\name{expr_binary}
\alias{expr_binary}
\alias{expr_grouping}
\alias{expr_literal}
\alias{expr_unary}
\title{Expressions}
\usage{
expr_binary(left, operator, right)

expr_grouping(expression)

expr_literal(value)

expr_unary(operator, right)
}
\arguments{
\item{left, right}{tokens}

\item{operator}{operator token}
}
\description{
Expressions
}
