% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashmap.r
\name{HashMap}
\alias{HashMap}
\title{HashMap}
\description{
HashMap

HashMap
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{keys}}{list of keys}

\item{\code{size}}{number of keys}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HashMap$new()}}
\item \href{#method-put}{\code{HashMap$put()}}
\item \href{#method-get}{\code{HashMap$get()}}
\item \href{#method-has}{\code{HashMap$has()}}
\item \href{#method-rm}{\code{HashMap$rm()}}
\item \href{#method-is_empty}{\code{HashMap$is_empty()}}
\item \href{#method-print}{\code{HashMap$print()}}
\item \href{#method-clone}{\code{HashMap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashMap$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-put"></a>}}
\if{latex}{\out{\hypertarget{method-put}{}}}
\subsection{Method \code{put()}}{
put add key/value to hashmap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashMap$put(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{any hashable R object}

\item{\code{value}}{any R object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
get value given key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashMap$get(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{any hashable R object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has"></a>}}
\if{latex}{\out{\hypertarget{method-has}{}}}
\subsection{Method \code{has()}}{
has check if key exists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashMap$has(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{any hashable R object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rm"></a>}}
\if{latex}{\out{\hypertarget{method-rm}{}}}
\subsection{Method \code{rm()}}{
rm remove key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashMap$rm(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{any hashable R object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_empty"></a>}}
\if{latex}{\out{\hypertarget{method-is_empty}{}}}
\subsection{Method \code{is_empty()}}{
is_empty check if hashmap is empty
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashMap$is_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print show hashmap
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashMap$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HashMap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
