% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret-expressions.r
\name{lookup_variable}
\alias{lookup_variable}
\title{Look-up variable in locals table}
\usage{
lookup_variable(name, x, interpreter)
}
\arguments{
\item{name}{token}

\item{x}{expression}

\item{interpreter}{interpreter}
}
\description{
Look-up variable in locals table
}
