% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{lox_function}
\alias{lox_function}
\title{Create Lox  Function}
\usage{
lox_function(declaration, closure)
}
\arguments{
\item{declaration}{stmt_function}

\item{closure}{env}
}
\description{
Create Lox  Function
}
