% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.r
\name{stmt_function}
\alias{stmt_function}
\title{Function Declaration}
\usage{
stmt_function(name, params, body)
}
\arguments{
\item{name}{token}

\item{params}{list of tokens}

\item{body}{list of statements}
}
\description{
Function Declaration
}
