% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.r
\name{stmt_if}
\alias{stmt_if}
\title{If Statement}
\usage{
stmt_if(condition, then_branch, else_branch = NULL)
}
\arguments{
\item{condition}{expression}

\item{then_branch}{statement}

\item{else_branch}{statement}
}
\description{
If Statement
}
