% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.r
\name{stmt_variable}
\alias{stmt_variable}
\title{Variable Statement}
\usage{
stmt_variable(name, initializer = NULL)
}
\arguments{
\item{name}{token}

\item{initializer}{expression}
}
\description{
Variable Statement
}
