% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.r
\name{expr_binary}
\alias{expr_binary}
\alias{expr_grouping}
\alias{expr_literal}
\alias{expr_logical}
\alias{expr_unary}
\alias{expr_variable}
\alias{expr_assignment}
\title{Expressions}
\usage{
expr_binary(left, operator, right)

expr_grouping(expression)

expr_literal(value)

expr_logical(left, operator, right)

expr_unary(operator, right)

expr_variable(name)

expr_assignment(name, value)
}
\arguments{
\item{left, right}{expression}

\item{operator}{token}

\item{value}{expression}

\item{name}{token}
}
\description{
Expressions
}
